requirements.txt
attrs==25.4.0
Automat==25.4.16
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
constantly==23.10.4
cryptography==46.0.5
cssselect==1.4.0
defusedxml==0.7.1
filelock==3.24.3
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
itemadapter==0.13.1
itemloaders==1.4.0
items==0.6.5
jmespath==1.1.0
lxml==6.0.2
nulltype==2.3.1
packaging==26.0
parsel==1.11.0
Protego==0.6.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
PyDispatcher==2.0.7
pyOpenSSL==25.3.0
queuelib==1.9.0
requests==2.32.5
requests-file==3.0.1
Scrapy==2.14.1
service-identity==24.2.0
tldextract==5.3.1
tqdm==4.67.3
Twisted==25.5.0
typing_extensions==4.15.0
urllib3==2.6.3
w3lib==2.4.0
zope.interface==8.2
